refactor(text-to-image): moves "/webAPI/Server" to "/Server"
