+++ src/features/TextToImage/Server/ServerConnectionError.ts
