+++ src/features/TextToImage/Server/ServerSpecificationError.ts
